+++ packages/audio_streamer/CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 1.1.0
+* Able ot stream on Android
+* The plugin will now record as soon as the permission dialog ends. 
 
-* TODO: Describe initial release.
+## 1.0.0
+* Able to stream audio PCM data on iOS.